pyproject.toml
[project]
name = "expenses-tracker"
version = "0.1.0"
description = "A script for saving expenses to sql database"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "psycopg>=3.2.9",
    "python-dotenv>=1.1.1"
]
